+++ backend/todolist/src/main/java/kr/codesquad/todolist/controller/CardDto.java
@@ -1,17 +1,18 @@
-package kr.codesquad.todolist.card;
+package kr.codesquad.todolist.controller;
 
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
+import kr.codesquad.todolist.domain.Card;
+import kr.codesquad.todolist.service.CardByStatus;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 public class CardDto {
 	@Data
