+++ backend/todolist/src/main/java/kr/codesquad/todolist/service/CardStatusNumber.java
@@ -1,5 +1,6 @@
-package kr.codesquad.todolist.card;
+package kr.codesquad.todolist.service;
 
+import kr.codesquad.todolist.domain.Card;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
